Update README.md
This repository contains two projects that I did during my BTech, in Optimization Methods in Engineering. It has two projects which have both single variable and multivariable maximization and minimization functions. The first project involves the Bounding Phase and Golden Section Search to optimize the search space for function optimization. The second project involves the use of both the Bounding Phase and Golden Section Search Methods for Line search in Powell's Conjugate Direction Search method, provided with an initial point and a search direction. Then the final optimised value and point are printed out. The value of the function and its corresponding iteration number gets stored in a CSV file simultaneously as the algorithm evaluates. it is then uploaded to a Python Script to plot the function value vs Iteration number, to get the Convergence Plots for different functions.